commit
actualizacion de clases
actualizacion de diagrama
actualizacion del Readme.md

+++ Fifa.Core/Habilidad.cs
@@ -2,7 +2,7 @@ namespace Fifa.Core
 {
     public class Habilidad
     {
-        public int idHabilidad { get; set; }
+        public sbyte idHabilidad { get; set; }
         public string habilidad { get; set; }
         public string descripcion { get; set; }
     }
