+++ Fifa.Core/Posicion.cs
@@ -2,7 +2,7 @@ namespace Fifa.Core
 {
     public class Posicion
     {
-        public int id { get; set; }
+        public sbyte id { get; set; }
         public string posicion { get; set; }
     }
 }
